@@ -15,47 +15,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0-alpha.0",
+    "@angular/animations": "^6.0.3",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
     "@types/codemirror": "0.0.56",
-    "codemirror": "^5.37.0",
-    "core-js": "^2.5.5",
-    "keycloak-angular": "^1.2.5",
+    "codemirror": "^5.38.0",
+    "core-js": "^2.5.7",
+    "keycloak-angular": "^4.0.0-beta.1",
     "ng2-ace-editor": "^0.3.7",
     "ng2-codemirror": "^1.1.3",
     "ng4-codemirror": "0.0.1",
-    "npm-check-updates": "^2.14.1",
-    "rxjs": "^5.5.9",
+    "npm-check-updates": "^2.14.2",
+    "rxjs": "^6.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/language-service": "^5.2.9",
-    "@types/jasmine": "^2.8.6",
+    "@angular-devkit/build-angular": "~0.6.5",
+    "@angular/cli": "6.0.5",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^9.6.4",
-    "codelyzer": "^4.2.1",
+    "@types/node": "^10.1.3",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.2",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^1.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
     "karma-junit-reporter": "^1.2.0",
-    "protractor": "~5.3.1",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "protractor": "^5.3.2",
+    "ts-node": "~6.0.5",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2"
   }
 }
